Created components and styles folders to organize better, also renamed App.css to index.css
