+++ src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import App from "./App";
 import "semantic-ui-css/semantic.min.css";
-import "./App.css";
+import "./styles/index.css";
 
 ReactDOM.render(
   <React.StrictMode>
